Change data type of ID

+++ models/materials.js
@@ -4,10 +4,10 @@ const materials = async (queryAsync) => {
     try {
         const materialsTable = await queryAsync(
             `CREATE TABLE IF NOT EXISTS materials(
-                id INT AUTO_INCREMENT PRIMARY KEY,
+                id BIGINT(64) AUTO_INCREMENT PRIMARY KEY,
                 name VARCHAR(255) NOT NULL,
                 manufacturer VARCHAR(255) NULL,
-                stocks INT NULL,
+                stocks BIGINT(64) NULL,
                 created_at TIMESTAMP
             )`
         );
